Orthrus: add capture run command + LaunchAgent plist for macOS deployment
- Add `orthrus capture run` command: starts CaptureManager as a foreground
  daemon with SIGTERM/SIGINT signal handling for clean shutdown.
  Blocks until signaled, then calls manager.shutdown(timeout=30s) and exits 0.
- Update LaunchAgent plist: entry point is now `orthrus capture run`
  (was `capture enable` which only set a config flag and exited).
  Add deploy/macos/com.orthrus.capture.plist to repo.
- deployment.md: fix plist XML snippet to match corrected entry point.
- Fix _fmt_bytes type: int → float (mypy caught: n/=1024 produces float).
- Ruff import sort fixes in sync/ (already correct on main, no-op on push).

+++ src/orthrus/cli/commands/sync.py
@@ -89,7 +89,7 @@ def cmd_sync(
         raise typer.Exit(code=1)
 
 
-def _fmt_bytes(n: int) -> str:
+def _fmt_bytes(n: float) -> str:
     """Format bytes as human-readable string."""
     for unit in ("B", "KB", "MB", "GB", "TB"):
         if abs(n) < 1024:
